+++ FindPanel/OgreTextFindComponents/OgreTextFindBranch.m
@@ -4,7 +4,7 @@
  *
  * Creation Date: Sep 26 2003
  * Author: Isao Sonobe <sonoisa@gmail.com>
- * Copyright: Copyright (c) 2003-2018 Isao Sonobe, All rights reserved.
+ * Copyright: Copyright (c) 2003-2020 Isao Sonobe, All rights reserved.
  * License: OgreKit License
  *
  * Encoding: UTF8
@@ -108,8 +108,8 @@
 }
 
 - (NSUInteger)numberOfDescendantsInSelection:(BOOL)inSelection
-{
-    NSUInteger          numberOfDescendants = 0;
+{ 
+    unsigned            numberOfDescendants = 0;
     NSEnumerator        *enumerator = (NSEnumerator*)[self componentEnumeratorInSelection:inSelection];
     OgreTextFindLeaf    *aChild;
