+++ FindPanel/OgreTextFindThread.m
@@ -4,7 +4,7 @@
  *
  * Creation Date: Sep 26 2003
  * Author: Isao Sonobe <sonoisa@gmail.com>
- * Copyright: Copyright (c) 2003-2018 Isao Sonobe, All rights reserved.
+ * Copyright: Copyright (c) 2003-2020 Isao Sonobe, All rights reserved.
  * License: OgreKit License
  *
  * Encoding: UTF8
@@ -467,7 +467,7 @@
 
 
 /* Protected methods */
-- (unsigned)numberOfMatches
+- (NSUInteger)numberOfMatches
 {
 	return _numberOfMatches;
 }
