+++ FindPanel/OgreTextFindThreads/OgreHighlightThread.m
@@ -4,7 +4,7 @@
  *
  * Creation Date: May 20 2004
  * Author: Isao Sonobe <sonoisa@gmail.com>
- * Copyright: Copyright (c) 2004-2018 Isao Sonobe, All rights reserved.
+ * Copyright: Copyright (c) 2004-2020 Isao Sonobe, All rights reserved.
  * License: OgreKit License
  *
  * Encoding: UTF8
@@ -65,7 +65,7 @@
         alpha: &alpha];
     
     numberOfGroups = [regex numberOfGroups];
-    NSUInteger  i;
+    unsigned    i;
     BOOL        simple = ([regex syntax] == OgreSimpleMatchingSyntax);
     double      dummy;
     
@@ -174,7 +174,7 @@
     cancelledMessagePlural      = OgreTextFinderLocalizedString(@"%d strings highlighted. (canceled, %.3fsec)");
     
     NSString    *message;
-    unsigned    count = [self numberOfMatches];
+    NSUInteger  count = [self numberOfMatches];
 	if ([self isTerminated]) {
 		if (count == 0) {
 			NSBeep();
