+++ FindPanel/OgreTextFindThreads/OgreReplaceAllThread.m
@@ -4,7 +4,7 @@
  *
  * Creation Date: May 20 2004
  * Author: Isao Sonobe <sonoisa@gmail.com>
- * Copyright: Copyright (c) 2004-2018 Isao Sonobe, All rights reserved.
+ * Copyright: Copyright (c) 2004-2020 Isao Sonobe, All rights reserved.
  * License: OgreKit License
  *
  * Encoding: UTF8
@@ -54,7 +54,7 @@
         return;
     }
     
-    NSUInteger    stringLength = [string length];
+    NSUInteger  stringLength = [string length];
     
     NSRange     selectedRange = [aLeaf selectedRange];
 	if (![self inSelection]) {
@@ -151,7 +151,7 @@
     cancelledMessagePlural      = OgreTextFinderLocalizedString(@"%d strings replaced. (canceled, %.3fsec)");
     
     NSString    *message;
-    unsigned    count = [self numberOfMatches];
+    NSUInteger  count = [self numberOfMatches];
 	if ([self isTerminated]) {
 		if (count == 0) {
 			NSBeep();
