+++ RegularExpression/OGRegularExpressionEnumeratorPrivate.m
@@ -4,7 +4,7 @@
  *
  * Creation Date: Sep 03 2003
  * Author: Isao Sonobe <sonoisa@gmail.com>
- * Copyright: Copyright (c) 2003-2018 Isao Sonobe, All rights reserved.
+ * Copyright: Copyright (c) 2003-2020 Isao Sonobe, All rights reserved.
  * License: OgreKit License
  *
  * Encoding: UTF8
@@ -114,7 +114,7 @@
 	_startLocation = location;
 }
 
-- (void)_setNumberOfMatches:(NSUInteger)aNumber
+- (void)_setNumberOfMatches:(unsigned)aNumber
 {
 	_numberOfMatches = aNumber;
 }
