+++ RegularExpression/OGRegularExpressionPrivate.h
@@ -4,7 +4,7 @@
  *
  * Creation Date: Sep 01 2003
  * Author: Isao Sonobe <sonoisa@gmail.com>
- * Copyright: Copyright (c) 2003-2018 Isao Sonobe, All rights reserved.
+ * Copyright: Copyright (c) 2003-2020 Isao Sonobe, All rights reserved.
  * License: OgreKit License
  *
  * Encoding: UTF8
@@ -25,6 +25,7 @@ typedef enum OgreKindOfCharacters {
 	OgreKindOfNormal
 } OgreKindOfCharacter;
 
+// 正規表現構文
 extern OnigSyntaxType  OgrePrivatePOSIXBasicSyntax;
 extern OnigSyntaxType  OgrePrivatePOSIXExtendedSyntax;
 extern OnigSyntaxType  OgrePrivateEmacsSyntax;
@@ -33,6 +34,7 @@ extern OnigSyntaxType  OgrePrivateGNURegexSyntax;
 extern OnigSyntaxType  OgrePrivateJavaSyntax;
 extern OnigSyntaxType  OgrePrivatePerlSyntax;
 extern OnigSyntaxType  OgrePrivateRubySyntax;
+
 
 @interface OGRegularExpression (Private)
 
@@ -63,7 +65,7 @@ extern OnigSyntaxType  OgrePrivateRubySyntax;
 // 名前がnameのgroup number
 // 存在しない名前の場合は-1を返す。
 // 同一の名前を持つ部分文字列が複数ある場合は-2を返す。
-- (int)groupIndexForName:(NSString*)name;
+- (NSInteger)groupIndexForName:(NSString*)name;
 // index番目の部分文字列の名前
 // 存在しない名前の場合は nil を返す。
 - (NSString*)nameForGroupIndex:(NSUInteger)index;
